digitalvocabulary/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
django==5.1.2
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-spectacular==0.27.2
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
packaging==24.1
    # via gunicorn
psycopg[binary]==3.2.3
    # via -r requirements.in
psycopg-binary==3.2.3
    # via psycopg
pyjwt==1.7.1
    # via djangorestframework-simplejwt
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via psycopg
tzdata==2024.2
    # via
    #   django
    #   psycopg
uritemplate==4.1.1
    # via drf-spectacular
